[PATCH] Fix get_debug_info() for looking for .debug_info section
Looks like there is a bug in makedumpfile 1.2.0 It seems that get_debug_info()
tries to look for .debug_info section, but it has a mistake when checking
section name: it should quit the loop when finding .debug_info section.

The bug doesnt affect the program's functionality though, but should be fixed.

Signed-off-by: Nguyen Anh Quynh <aquynh@gmail.com>