[v1.2.7-04] Add the progress bar for creating dump bitmap.
On large nodes, bitmap creation takes a serious amount of time. But the
progress indicator only starts after the bitmaps has been created, showing
only the progress of the copy process.

This patch adds a "message" parameter to the print_progress() function.
That message is displayed first and gets overwritten by next message.
So, the user sees

    Excluding zero pages : [  0 %]
    Excluding zero pages : [... %]
    Excluding zero pages : [100 %]
    Copying data         : [  0 %]
    Copying data         : [... %]
    Copying data         : [100 %]

in sequence. IMO that's the fastest option to implement such a progress.


Signed-off-by: Bernhard Walle <bwalle@suse.de>
Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>