[v1.2.7-08] Shrink the time for creating dump bitmap.
The method for creating an ELF dump file wastes the time by excluding
zero pages, because it reads all each page and checks it. This checking
method is very heavy, and this patch shrinks this time by scanning only
dumpable pages.

The larger dump_level (-d option) is, more effective this patch is.
If dump_level is 1, it does not shrink the time.

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>