[v1.2.9-14] Add "--vtop" option for debugging.
This patch adds "--vtop" option to makedumpfile's command line.
This option is useful, when user debugs the translation problem
of virtual address. It makes debugging easy by comparing the
output of this option with the one of "vtop" subcommand of the
crash utility.

"--vtop" option only prints the translation output, and it does
not affect the dumpfile creation.


Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>