[v1.3.0-08] Fix the buffer handling problem for reading struct page.
from Dave Anderson:

In exclude_unnecessary_pages(), if a physical page is in a memory
hole, then it skips the page and continues.  In our case, that happened,
but when it started up again, the next legitimate pfn was well beyond
the previously-read cache of 512 pages.  But since the new legit page
wasn't modulo-512, it didn't refresh the page cache, and it ended up
using stale page data (page->flags) and ended up excluding legitimate
page.

* Reference
  http://lists.infradead.org/pipermail/kexec/2008-September/002549.html

Tested-by: Jay Lan <jlan@sgi.com>
Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>