[v1.3.3-11] Fix vmcoreinfo fclose().
If specifying a vmcoreinfo file with -i option and /proc/vmcore contains
vmcoreinfo data, makedumpfile does not fclose() a vmcoreinfo file. Because
the file descriptor of a vmcoreinfo file is overwritten to the one of
vmcoreinfo data in /proc/vmcore before fclose().

The following is the problematic sequence:
 1. Specify a vmcoreinfo file by -i option.
    (ex. This vmcoreinfo file name is VMCOREINFO)
 2. fopen() VMCOREINFO in open_vmcoreinfo().
 3. Get the debug information from VMCOREINFO.
 4. Overwrite the file name of a vmcoreinfo file by a tempolary file.
   (ex. This tempolary file name is VMCOREINFO_IN_VMCORE)
 5. fclose() VMCOREINFO_IN_VMCORE in close_vmcoreinfo().

So makedumpfile does not close VMCOREINFO file.

This patch fixes this leak problem. If applying this patch, vmcoreinfo
file is closed immediately just after getting the debug information from it.

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>