[v1.3.3-12] Use current page_size for KDUMP files.
If the kdump kernel has a different page size than the crashed kernel,
the wrong page size gets used when computing the size of the KDUMP
page bitmap even if we have a vmcoreinfo.

Move the calculation beyond processing vmcoreinfo, and issue a warning
if the default page size gets used.


Signed-off-by: Petr Tesarik <ptesarik@suse.cz>
Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>