[v1.3.4-10] Add dump filtering on an x86_64 xen domain-0.
This patch adds the dump filtering for excluding unnecessary pages (cache
pages, user process data pages, and free pages) on on x86_64 xen domain-0.

On the existing makedumpfile (v1.3.3 or former), a user could specify 0
or 1 only as a dump_level. By this patch, he/she can specify 2 or larger
also as a dump_level.

Now, this feature is effective on x86_64 machine only.


Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>