[v1.3.4-13] Support kdump-compressed format for input file.
This is a patch for improving makedumpfile so that it can read dumpfile
which is kdump-compressed format and re-filter it.

For example,

  makedumpfile -d 3 -c /proc/vmcore vmcore.1

After this, if user wants to reduce the size of vmcore.1, for example,
if they need to reduce the size because they want to send it somewhere
using ftp/scp/etc, this feature is useful.

  makedumpfile -d 31 -c vmcore.1 vmcore.2

We can re-filter vmcore.1 file and reduce the size using this feature.

[How this patch work?]
After this patch is applied, makedumpfile works as follows.
 1) If input file is kdump-compressed, read a dump header.
 2) When 1st-bitmap is generated, just copy it from the kdump-compressed
    file to the output file.
 3) When makedumpfile reads memory, readpmem_kdump_compressed() is used
    instead of readmem(). readpmem_kdump_compressed() is new function I
    added, it just reads data from kdump-compressed file.


Signed-off-by: Takao Indoh <tindoh@redhat.com>
Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>