[v1.3.4-14] Add vaddr_to_paddr_ppc64().
If input file is kdump-compressed format, vaddr_to_paddr_general
does not work. As to x86/x86_64/ia64, there is a function to translate
address and it can be used instead of vaddr_to_paddr_general, but there
is not on ppc. This patch adds new translation function
(vaddr_to_paddr_ppc64) for ppc.


Signed-off-by: Takao Indoh <tindoh@redhat.com>