[PATCH] Cleanup: the initialization method is operated if it is possible.
On current code, the initialization method is operated only if
the method is necessary. That is not wrong, but that is a little
complex because makedumpfile has many options and there are many
situations the method is necessary or unnecessary.
So the initialization method is operated always if it is possible.

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>