makedumpfile: output files filling ramdisk
Tested on SGI UV, which is x86_64.

makedumpfile's path FILENAME_BITMAP "/tmp/..." causes it to try to write
bitmap files to the ramdisk.  These can be very large if the machine has
lots of memory, and so writing the file can fail.

(makedumpfile is exec'd by kdumptool. And kdumptool should chroot to
 its --root directory.  But if the intent is to avoid this problem it
 doesn't seem to work.  That may be the proper fix, rather than the below.)

A TMPDIR environment variable is set in the script
/lib/mkinitrd/scripts/boot-kdump.sh. So this patch assumes that
the FILENAME_BITMAP path should be based at TMPDIR.

Diffed against makedumpfile-1.3.5

Signed-off-by: Cliff Wickman <cpw@sgi.com>
Signed-off-by: Masayuki Igawa <igawa@mxs.nes.nec.co.jp>