[PATCH] Add ELF note section to the kdump-compressed format.
In case of s390x's DUMPFILE in the kdump-compressed format, the crash
utility gets the register information using CPU lowcore symbol.
All s390 dumps have the register information in memory on the lowcore
pages.

But in the future s390 could have a dump mechanism that does not store
the register information in the lowcore pages in the memory and will
only keep it on ELF note section.

So DUMPFILE in the kdump-compressed format needs ELF note section for
including the register information, and this patch adds new members
"offset_note" and "size_note" into kdump sub header for representing
the section.

Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>