[PATCH v2 5/8] Read and process filter commands from config file.
This patch enables to makedumpfile to read and process filter commands from
specified config file. It builds a list of filter info consisting of memory
address (paddr) and size to be filtered out. This list is used during
read_pfn() function to filter out user specified kernel data from vmcore
before writing it to compressed DUMPFILE. The filtered memory locations are
filled with 'X' (0x58) character.

The filter command syntax is:
	erase <Symbol>[.member[...]] [size <SizeValue>[K|M]]
	erase <Symbol>[.member[...]] [size <SizeSymbol>]
	erase <Symbol>[.member[...]] [nullify]

Below are the examples how filter commands in config file look like:

	erase modules
	erase cred_jar.name size 10
	erase vmlist.addr nullify

This is useful feature for the customers who wants to erase the customer
sensitive data like security keys, in DUMPFILE before sending it to
support team for analysis.

The man page for makedumpfile.conf(8) has been introduced that describes the
config file format and erase command syntax.

Change in v2:
- Renamed filter.conf file to makedumpfile.conf
- Added cleanup routine clear_filter_info() to freeup memory allocated during
  gather_filter_info().
- Modified gather_filter_info() function to return error if
  process_config_file() can't open configuration file.

Signed-off-by: Mahesh Salgaonkar <mahesh@linux.vnet.ibm.com>
Signed-off-by: Prerna Saxena <prerna@linux.vnet.ibm.com>
---