[PATCH] Support line mode terminals.
On S390 we have line mode terminals that do not have the possibility to
position the cursor. When using kdump the output of makedumpfile will be
shown on such terminals. Currently the output looks very strange because
the '\r' control character is ignored. It looks like the following:

Checking for memory holes          : [  0 %]  Checking for memory holes
: [100 %]  Copying data                       : [ 33 %]  Copying data
            : [ 65 %]  Copying data                       : [ 96 %]
Copying data                       : [100 %]

For those line mode terminals the TERM environment variable is set to
"dumb".
With this patch makedumpfile checks the TERM variable and in case of
"dumb" instead of using '\r' each progress message is printed into a
separate line.
The output will then look like the following:

    Checking for memory holes          : [  0 %]
    Checking for memory holes          : [100 %]
    Copying data                       : [ 11 %]
    Copying data                       : [ 44 %]
    Copying data                       : [ 75 %]
    Copying data                       : [100 %]

Signed-off-by: Michael Holzheu <holzheu@linux.vnet.ibm.com>
Signed-off-by: Ken'ichi Ohmichi <oomichi@mxs.nes.nec.co.jp>