[PATCH] sadump: Check if given cpu is online in per-cpu related helper functions.
Per-cpu helper functions sanity check given cpu ids if they are
not-minus and less than the maximal but it should have checked if they
are online. By this mistake, one cannot see generated vmcore correctly
if sadump collects vmcore when some cpus are offline e.g. using cpu
hotplug or panic and so all the cpus except for a unique CPU halting.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>