[PATCH v4 9/12] Enabling --split option with cyclic mode.
This patch enable the cyclic mode to split the dump data to multiple
dumpfiles. At this time, the buffer of partial bitmap is prepared for
each child process, you need to consider the amount of memory usage.

This patch split the dump data based on only max_mapnr and num_dumpfile,
so the size of each split dumpfiles will be different by excluding
unnecessary pages.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>