[PATCH] Support for x86_64 1G pages.
The PS bit is not recognized in the Page-Directory-Pointer Table (pgdir
in Linux terms), so virtual addresses that map to a 1G page will be
translated incorrectly.

This bug affects both Xen and bare metal. I don't think it can be easily
triggered in practice, because 1G mappings are created only for:

1. direct 1:1 mapping of physical memory, for which we don't walk the page
   tables, but instead subtract the corresponding virtual offset, and
2. 1G hugepages, which are only used for userspace data.

Anyway, if we ever happen to hit a 1G page now, the virtual address will
be translated correctly.

While fixing this, I consolidated the vtop() translation routines and
changed some of the constants. The basic idea is that the format of a page
table entry is defined by the architecture, so we can use hard-coded
constants everywhere.

I also always mask off reserved bits, because they might get defined in
a later revision of the specification.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>