[PATCH] keep dumpfile pages in a cache.
Add a simple cache for pages read from the dumpfile.

This is a big win if we read consecutive data from one page, e.g.
page descriptors, or even page table entries.

Note that makedumpfile now always reads a complete page. This was already
the case with kdump-compressed and sadump formats, but makedumpfile was
throwing most of the data away. For the kdump-compressed case, we may
actually save a lot of decompression, too.

I tried to keep the cache small to minimize memory footprint, but it should
be big enough to hold all pages to do 4-level paging plus some data. This
is needed e.g. for vmalloc areas or Xen page frame table data, which are not
contiguous in physical memory.

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>