[PATCH v3 4/9] Split Xen setup into early and late.
Previously, get_xen_info_arch() was called after initializing the generic
Xen variables. This is necessary, because the arch-specific code translates
domain addresses to pickled domain IDs, and the domain list is initialized
by the generic code.

Now, the virtual memory layout on x86_64 was changed in Xen4, and this
initialization should ideally be done in the arch-specific code.
Unfortunately, the generic code must know the virtual address space to
perform readmem of type XEN_VADDR, so there is no working ordering, and
instead we must provide hooks for "early" arch-specific initialization
to set up data needed by the generic code, and "late" arch-specific
initialization to do stuff that depends on data set by the generic code
(currently only the domain_list).

Signed-off-by: Petr Tesarik <ptesarik@suse.cz>