[PATCH v3 9/9] Do not fail for symbols removed in Xen4.
Xen-4.0+ removes some of the symbols that were necessary to process dump
files on Xen-3, and makedumpfile fails if they are not found.

Since most of these symbols are not needed for Xen4, we can simply remove
the checks. The only exception is "frame_table". This used to be a global
variable holding the address of the array of struct page_info. In Xen-4.0
the array is always located at a fixed address in virtual address space.
Xen adds a compatibility symbol for VMCOREINFO, but a "frame_table" symbol
does not exist as such, so we must use the fixed address instead.

Signed-off-by: Norbert Trapp <norbert.trapp@ts.fujitsu.com>
Signed-off-by: Petr Tesarik <ptesarik@suse.cz>