[PATCH v2 09/10] Add page_is_buddy for old kernels.
On kernels from v2.6.15 to v2.6.17, buddy page is marked by the
condition that PG_private flag is set and _count == 0.

Since this page_is_buddy needs _count, we use freelist logic if _count
is not available.

Unfortunately, I have yet to test this logic on these kernel versions
simply because I've been failing to boot them on my box.

Note that on these kernels, free list can be corrupted due to the bug
that the above two conditions are not checked atomically. The reason
why PG_buddy was introduced is a fix for this bug.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>