[PATCH] sadump: fix warning messages in building time.
Recent gcc outputs the following warning messages when compiles
sadump_info.c because a given pointer is compared as conditional but
they are static string objects and are alreays converted into true.

./sadump_info.c: In function 'sadump_virt_phys_base':
./sadump_info.c:926: warning: the address of 'buf' will always evaluate as 'true'
./sadump_info.c:933: warning: the address of 'buf' will always evaluate as 'true'

But this is within our intension so enclose the occurence of static
strings by static inline function to surpress the compiler's checking.

I borrowed this code from crash utility.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>