[PATCH] Clean up readmem() by removing its recursive call.
Currently, readmem() calls itself recursively and its depth grows in
proportion to the number of requested pages.

For example, in __exclude_unnecessary_pages(), PGMM_CACHED is defined
as 512 and page structure is of 56 bytes on x86_64, so size of page
descriper cache on x86_64 is now 56 * 512 = 28KiB, and this is 7
pages. This means that on the existing implementation, readmem() is
called recursively 7 times when reading one entry of page descripter
cache from mem_map array, but 6 of the 7 are in fact unnecessary.

This patch cleans up readmem() by removing the recursive
call. Instead, readmem() proceeds to the processing for next page by
jump.

By this change, read operation is done in increasing order. This is
more natural than the existing implementation in decreasing order.

Also, there's visible performance gain. On 32GiB memory machine,
repeating filtering 100-times on /proc/vmcore, average time for
filtering on mem_map array was improved from 1.69 to 1.33 second:
about 21.3 percent.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>