[PATCH] Fix max offset relative to file.
To see file offset of each memory chunk, it's correct to read
p_offset in the corresponing PT_LOAD entrie.

On /proc/vmcore PT_LOAD entries are sorted on p_load values in
increasing order. So, it's sufficient to refer to the last PT_LOAD
entry only. But the code here doesn't assuming that, calculating
maximum one from all the PT_LOAD entries.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>