[PATCH] Fix missing count of filtered free pages.
On the current implementation, __exclude_unnecessary_pages() can refer
to page frames not belonging to current cycle. Thus, pfn_free can be
now counted up even if it doesn't belong to the current cycle.

This patch changes it so pfn_free is counted up only if the
corresponding pfn really belongs to the cycle.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>