[PATCH] Skip multiple pages filtered as free pages.
If buddy page has order n, (2^n) pages are filtered at the same
time. Then, next (2^n) pages are already filterd, so we can skip the
(2^n) pages for free.

By this, on 32GB memory system, there's the following improvement:

- with mmap() on /proc/vmcore from 0.594421 seconds to 0.496517
  seconds, and
- with normal read() from 1.533796 seconds to 0.750935 seconds.

Note that the two improvements are measured on the same /proc/vmcore.

Note also that amount of the improvement depends on amount of free
pages, though it's obvious.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@jp.fujitsu.com>