[PATCH 2/2] Fix the issue which can happen around overlapping segments.
If reading the separated page on different PT_LOAD segments,
we have to read the page data from both segments and read_pfn()
was introduced for such cases.
(Please see the comment of read_pfn() for details.)

But read_pfn() hasn't worked as expected since the caching
feature was introduced.
The current call stack can be broken into as follows:

  read_pfn()
    readmem()
      readpage_elf()
        read(2)

If the target of read_pfn() is separated page, readmem() is called
twice as for the former part and later part of the page.
But readpage_elf() always try to read a whole page and cache it
without consideration for separated page. As the result, readmem()
can get the same invalid data in both readmem().

So, the code for separated page should be moved from read_pfn()
to readpage_elf().

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>