[PATCH] Check dump file early.
On s390 the makedumpfile tool sometimes is used directly by
users on the command line. Currently the check if the dump
file already exists is done after the filtering function has
been called. Therefore, for large dumps the user has to wait
for filtering and after some time he gets the error message
"open_dump_file: Can't open the dump file(out). File exists".

This patch improves this by adding an early check for the
existence of the dump file. In case the -f (force) option has
been specified it is checked that an existing file is writable.

Signed-off-by: Michael Holzheu <holzheu@linux.vnet.ibm.com>