[PATCH v2] "make install" assumes directories exist.
The make install of makedumpfile-1.5.4 uses cp instead of the usual
"install" command, which has some side effects of not pre-creating
directories.  This results in installs to an alternate DESTDIR fail
because the parent directories do not exist.

I have attached a patch which modifies make install to use the install
utility.

Signed-off-by: Nick Bartos <nick@pistoncloud.com>

Change installation directory to /usr/sbin.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>