[PATCH] Restrict the mmap range to avoid mmap() failure.
Now, we know that mmap() on /proc/vmcore for fractional pages
will fail due to a sanity check in remap_pfn_range().
In that case, current makedumpfile can fall back to read() mode,
but it's still desirable to use mmap() as much as possible for
performance.
This patch restricts the range of mmap() to avoid the fall back.

The original kernel issue was reported in:

  https://lkml.org/lkml/2013/11/13/439

and it will be fixed in kernel 3.14.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>