[PATCH] eppic: Add support to module data structures.
When trying to refer a module data structure GET_DIE_NFIELDS
does not set the dwarf_info to the right module.

For instance:

struct ap_device {
	...
	struct list_head list;
	...
};

struct ap_device *dev;

As a result, it was able to resolve only dev->list and it was failing to
resolve dev->list.next.
The patch takes care of handling this by introducing GET_DIE_NFIELDS_ALL
which takes care of setting the dwarf_info to the appropriate module.

Signed-off-by: Aruna Balakrishnaiah <aruna@linux.vnet.ibm.com>