[PATCH] Add a warning to detect wrong cycle iteration.
set_bitmap_cyclic() has a sanity check to reject pfn which is out of
the current cycle range, but we don't pass such pfn to set_bitmap_cyclic()
in principle. So if the check fails, it's an unexpected behavior for us,
but we can never know that fact now.

This patch add a warning to know that, it will appear only once at
runtime if the caller side logic has bugs.

Signed-off-by: Atsushi Kumagai <kumagai-atsushi@mxc.nes.nec.co.jp>