[PATCH] Fix bug on is_sparsemem_extreme.
The kernel gets the length of mem_section use this
from: include/linux/mmzone.h:1134

 "#define NR_SECTION_ROOTS  DIV_ROUND_UP(NR_MEM_SECTIONS,
					 SECTIONS_PER_ROOT)"

But when we check the memory model. We forget the
"DIV_ROUND_UP" method.

So when 32-bit ARM linux is in sparsemem_extreme memory:
 with SECTION_SIZE_BIT=28 and 16 byte-size mem_section.
The makedumpfile can not  identify the memory correctly.

This patch introduce the "DIV_ROUND_UP" method in this check.

Signed-off-by: Liu Hua <sdu.liu@huawei.com>